Fix printing size_t and off_t value for LLP64 Windows.
* src/debug.h (_trace_sysres_off_t): Use 'z' length specifier.
* src/w32-io.c (_gpgme_io_read): Likewise.
(writer, _gpgme_io_write): Likewise.

--

GnuPG-bug-id: 7187

+++ src/debug.h
@@ -164,7 +164,7 @@ static inline gpgme_off_t
 _trace_sysres_off_t (gpgme_off_t res, int lvl, const char *func, int line)
 {
   if (res >= 0)
-    _gpgme_debug (NULL, lvl, 3, func, NULL, NULL, "result=%ld", res);
+    _gpgme_debug (NULL, lvl, 3, func, NULL, NULL, "result=%zd", (size_t)res);
   else
     _gpgme_debug (NULL, lvl, -1, NULL, NULL, NULL,
                   "%s:%d: error: %s (%d)\n",
